backend/requirements.txt
numpy==2.3.5
openai==2.8.0
openai-agents==0.6.1
pandas==2.3.3
python-dotenv==1.2.1
pypdf
fastapi==0.121.3
uvicorn[standard]==0.38.0
requests==2.32.5
httpx==0.28.1
youtube-transcript-api==1.2.3
supabase==2.24.0
langfuse==3.10.5
openinference-instrumentation-openai-agents
